move auth package into "cf" package
Authored-by: Connor Braa <cbraa@pivotal.io>
