+++ src/cf-api-kpack-watcher/config/crd/build.yaml
@@ -1,3 +1,4 @@
+# this is copied from kpack's release and is used in test
 ---
 apiVersion: apiextensions.k8s.io/v1beta1
 kind: CustomResourceDefinition
